json loader in C# (+basic unit test)

+++ src/examples/CrazyRays/GroundWrapper/Shading/Spectrum.cs
@@ -23,7 +23,7 @@
         public static ColorRGB White =
             new ColorRGB { r = 1.0f, g = 1.0f, b = 1.0f };
 
-        public ColorRGB(int r, int g, int b) : this() {
+        public ColorRGB(float r, float g, float b) : this() {
             this.r = r;
             this.g = g;
             this.b = b;
